Split PackedBits/MutablePackedBits into their own packed_bits.h
Move the generic packed-bits view types out of quantized_vector.h into a
dedicated packed_bits.h. This keeps the layering clean: the generic
MultiBitPacker now depends only on packed_bits.h instead of the
EDEN-specific quantized_vector.h, while quantized_vector.h reuses the
same PackedBits type by including the new header.
